requirements.txt
Django>=5.2.4
mysqlclient>=2.2.0
django-cors-headers>=4.3.1
django-rest-framework>=3.15.1
python-dotenv>=1.0.1
